fix: remove off-diagonal contributions when initialising f10 in vfp1d (#126)
Previously all tests were skipped but the intention seemed to be that they ran for 1, 20, 80.

These failed with ee True because the matrix later on during the sim was ill-posed. This was due to the off-diagonal terms added explicitly, added a flag to remove them on initialisation. Probably best to put a switch on whether initialisation is desired in the future but hopefully this fixes for now.

Tests also failed due equinox + Jax 0.7.1, pinned to 0.6.2 to solve.

Also separated tests to run in parallel we might want to add checks about whether a particular folder was updated when deciding whether to run in the future.